test(nilcomm): tidy test channel queue binding & message purging
