+++ vitest.config.ts
@@ -5,7 +5,6 @@ export default defineConfig({
   plugins: [tsconfigPaths()],
   test: {
     testTimeout: 0,
-    // @ts-expect-error ts not recognising reporter key as per https://vitest.dev/guide/coverage.html#custom-coverage-reporter
     coverage: {
       reporter: ["text", "json-summary", "json"],
       reportOnFailure: true,
